fix: time and activity report dates displayed issue for chart & grid
