+++ apps/gauzy/src/app/pages/reports/time-reports/time-reports/time-reports.component.html
@@ -41,10 +41,14 @@
 		<div class="daily-time-report">
 			<nb-card>
 				<nb-card-body>
-					<ngx-line-chart [data]="chartData"></ngx-line-chart>
+					<ngx-line-chart 
+						[data]="chartData"
+					></ngx-line-chart>
 				</nb-card-body>
 			</nb-card>
 		</div>
-		<ga-daily-grid [filters]="filters"></ga-daily-grid>
+		<ga-daily-grid 
+			[filters]="filters"
+		></ga-daily-grid>
 	</nb-card-body>
 </nb-card>
